Ajout de la liste des remboursements

+++ src/app/refunds/mutations/deleteRefund.ts
@@ -7,7 +7,14 @@ export default resolver.pipe(
   resolver.authorize(),
   async ({ id }) => {
     // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-    const refund = await db.refund.deleteMany({ where: { id } })
+
+    const refund = await db.refund.findFirst({ where: { id } })
+
+    if (refund?.isValidated === true) {
+      throw new Error("Vous ne pouvez pas supprimer un remboursement validé !")
+    }
+
+    await db.refund.deleteMany({ where: { id } })
 
     return refund
   }
